refactor(core): add resource instrumentation hook
