+++ docs/api-reference/core/device.md
@@ -78,6 +78,7 @@ Specifies props to use when luma creates the device.
 | `onPositionChange?: (ctx: CanvasContext) => unknown`    | `log.info(1)`                                | Called if the canvas position changes. Requires `CanvasContextProps.trackPosition` to be enabled.           |
 | `onVisibilityChange?: (ctx: CanvasContext)`             | `log.info(1)`                                | Called if the visibility of the canvas changes (window is closed or occluded).                              |
 | `onDevicePixelRatioChange?: (ctx: CanvasContext)`       | `log.info(1)`                                | Called if the DPR changes (perhaps by moving the window to another screen                                   |
+| `resourceInstrumentation?: ResourceInstrumentation \| null` | default stats instrumentation                 | Observes resource lifecycle and allocation events. Set to `null` to disable resource stats bookkeeping.     |
 | `debug?`: `boolean`                                     | `false`                                      | Extra checks (wait for shader compilation, framebuffer completion, WebGL API errors will throw exceptions). |
 | `debugGPUTime?: boolean`                                | `false`                                      | Enable GPU timestamp collection without enabling all debug validation paths.                                 |
 | `debugShaders?`: `'errors' 'warnings' 'always' 'never'` | `'error'`                                    | Display shader source code with inline errors in the canvas.                                                |
